+++ crystal/__init__.py
@@ -1,5 +1,5 @@
 __version__ = "1.0.0"
-__build__ = "003"
-__release__ = "Crystal V1.00 Build 003"
+__build__ = "004"
+__release__ = "Crystal V1.00 Build 004"
 
 __all__ = ["__version__", "__build__", "__release__"]